NANOGrav TWG WB noise setup
Introduce the setup that the NANOGrav TWG uses for the wideband datasets. The only change is to place normal priors on DMEFAC and EFAC.

This introduces two flags, `ng_twg_setup = False` and `wb_efac_sigma = 0.25`, which toggle on this mode and set the sigma for the normal prior on EFAC parameters respectively. There's also an assoicated change in `blocks.py` to allow for normal prior on the EFAC.