fix(2D KDE): pin scipy
SciPy has deprecated interp2d used in the 2D KDE method.
Their suggested backwards compatible method is not actually backwards
compatible because it doesn't expose the same arguments to users, like
the fill value outside of the domain. The suggested new methods take
different inputs, so it is unclear if they will work with the empirical
distribution code. This needs to be further tested before changing interp2d.